+++ Assets/Mirage/Components/AssemblyInfo.cs
@@ -1,3 +1,6 @@
-﻿using System.Reflection;
+using System.Reflection;
+using System.Runtime.CompilerServices;
 
-[assembly: AssemblyVersion("89.0.0")]
+[assembly: AssemblyVersion("92.0.0")]
+
+[assembly: InternalsVisibleTo("Mirage.Tests.Runtime")]