+++ Assets/Mirage/Components/NetworkManagerHud.cs
@@ -41,14 +41,14 @@ namespace Mirage
         public void StartHostButtonHandler()
         {
             SetLabel("Host Mode");
-            NetworkManager.Server.StartHost(NetworkManager.Client).Forget();
+            NetworkManager.Server.StartAsync(NetworkManager.Client).Forget();
             OnlineSetActive();
         }
 
         public void StartServerOnlyButtonHandler()
         {
             SetLabel("Server Mode");
-            NetworkManager.Server.ListenAsync().Forget();
+            NetworkManager.Server.StartAsync().Forget();
             OnlineSetActive();
         }
 
@@ -62,7 +62,7 @@ namespace Mirage
         public void StopButtonHandler()
         {
             SetLabel(string.Empty);
-            NetworkManager.Server.StopHost();
+            NetworkManager.Server.Stop();
             NetworkManager.Client.Disconnect();
             OfflineSetActive();
         }