+++ Assets/Tests/SocketLayer/PeerTestAsClient.cs
@@ -11,14 +11,6 @@ namespace Mirage.SocketLayer.Tests.PeerTests
     [Category("SocketLayer"), Description("tests for Peer that only apply to client")]
     public class PeerTestAsClient : PeerTestBase
     {
-        readonly byte[] connectRequest = new byte[3]
-          {
-            (byte)PacketType.Command,
-            (byte)Commands.ConnectRequest,
-            new ConnectKeyValidator().GetKey(),
-          };
-
-
         [Test]
         public void ConnectShouldSendMessageToSocket()
         {