+++ Assets/Tests/SocketLayer/PeerTestConnecting.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
+using Mirage.Tests;
 using NSubstitute;
 using NUnit.Framework;
 
@@ -22,10 +23,6 @@ namespace Mirage.SocketLayer.Tests.PeerTests
             for (int i = 0; i < ClientCount; i++)
             {
                 clients[i] = new PeerInstanceWithSocket();
-
-                // add remotes so they can send message to each other
-                server.socket.AddRemote(clients[i].socket);
-                clients[i].socket.AddRemote(server.socket);
             }
         }