+++ Assets/Tests/Weaver/GeneratedReaderWriterTests~/CreatesForNullable.cs
@@ -1,0 +1,11 @@
+using Mirage;
+using UnityEngine;
+
+namespace GeneratedReaderWriter.CreatesForNullable
+{
+    [NetworkMessage]public struct NullableIntMessage { public int? value; }
+    [NetworkMessage]public struct NullableBoolMessage { public bool? value; }
+    [NetworkMessage]public struct NullableUlongMessage { public ulong? value; }
+    [NetworkMessage]public struct NullableUshortMessage { public ushort? value; }
+    [NetworkMessage]public struct NullableVector3Message { public Vector3? value; }
+}